+++ fraud/src/main/java/com/lugro/FraudCheckService.java
@@ -10,7 +10,7 @@ import lombok.AllArgsConstructor;
 @AllArgsConstructor
 public class FraudCheckService {
     private final FraudCheckHistoryRepository fraudCheckHistoryRepository;
-    public boolean isFraudulentCustomer(Integer customerId) {
+    public boolean isFraudulentCustomer(Long customerId) {
         fraudCheckHistoryRepository.save(
             FraudCheckHistory.builder()
                 .customerId(customerId)
